Update: Simplify logic
